# api-swagger-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ECommerce Application
  description: >-
    This is E-commmerce REST API which provides simple endpoints which can be used by
    anyone to instantly to lauch a backend Ecommerece servers with minimum
    customizations and also helps to scale within less time
    
    - Project Details :
    
        - [Backend Server Code](https://github.com/Vinay-M-G/Simple-Commerece-Backend)
        - [Frontend Application Code](https://github.com/Vinay-M-G/Simple-Commerece--frontend-)
  contact:
    email: vinay.giriswamy@gmail.com
  version: 3.0.0
servers:
  - url: http://{EC2 Server Id}:9001/api/v1
    variables:
      EC2 Server Id:
        default: Add_domain_here
...
paths:
...
  /carts/basketproductupdate/:
    patch:
      tags:
        - Carts Controller
      summary: Updates product quantity 
      description: |-
        - Adds specified quantity of product
        - Send 0 to remove product from basket
      parameters:
        - name: productCode
          in: query
          description: provide 9 digit product code
          required: true
          schema:
            type: string
        - name: quantity
          in: query
          description: provide quantity of product
          required: true
          schema:
            type: integer
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel'
        '409':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel'        
  /carts/basketserviceupdate/:
    patch:
      tags:
        - Carts Controller
      summary: Adds product service to basket for the specified product
      description: |-
        - Adds specified service for the required product
        - Send isSelected boolean as true to add product service to basket
        - Send isSelected boolean as false to remove product service from basket
      parameters:
        - name: productCode
          in: query
          description: provide 9 digit product code
          required: true
          schema:
            type: string
        - name: servicecode
          in: query
          description: provide service Id associated with the product
          required: true
          schema:
            type: string
        - name: isSelected
          in: query
          description: Boolean which tells whether the service to be removed or added for the praticular product
          required: true
          schema:
            type: boolean    
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel' 
        '409':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel'
  /carts/addtobasket/:
    post:
      tags:
        - Carts Controller
      summary: Adds product to basket
      description: |-
        - Adds single quantity of product to basket
      parameters:
        - name: productCode
          in: query
          description: provide 9 digit product code
          required: true
          schema:
            type: string
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel'
        '409':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasketResponseModel'
...
  /carts/loadbasket/:
    get:
      tags:
        - Carts Controller
      summary: provides details of user cart
      description: |-
        - provides details of user cart
      parameters:
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCartModel'
        '404':
          description: cart not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  guid:
                    type: string
  /orders/getorder/{orderId}:
    get:
      tags:
        - Order Controller
      summary: provides order details based on order id
      description: |-
        - provides details of an order based on order id
      parameters:
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
        - name: orderId
          in: path
          description: Order id in format xxx-xxxx-xxx. 
          required: true
          schema:
            type: string    
      responses:
        '200':
          description: Successfull response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderModel'
        '404':
          description: Order Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
  /orders/createorder/{cartPrice}:
    post:
      tags:
        - Order Controller
      summary: Creates an Order 
      description: |-
        - Creates an order if there are no validation errors
        - Cart will be validated based on parameters such as availability of stock, online sellability of the product, changes in cart price, valid guid
        - Once validation is succeeded, order will be saved and a new order id will be created
      parameters:
        - name: guid
          in: header
          description: provide guid of the basket to identify basket
          required: true
          schema:
            type: string
        - name: cartPrice
          in: path
          description: provide the final payable price of the cart. Should be greater than 1. 
          required: true
          schema:
            type: string    
      responses:
        '201':
          description: Successfull response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                  orderId:
                    type: string
        '404':
          description: Cart Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '409':
          description: Validation Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'        
...
components:
  schemas:
...
    BasketResponseModel:
      type: object
      properties:
        error:
          type: boolean
        message:
          type: string
...
    ServiceModel:
      type: object
      properties:
        serviceType:
          type: string
          enum:
            - Product
            - Delivery
        priceValues:
          type: object
          properties:
            basePrice:
              type: number
            offerPrice:
              type: number
        isMultiplier:
          type: boolean
        serviceDescription:
          type: string
        serviceName:
          type: string
        serviceId:
          type: string
    UserCartModel:
      type: object
      properties:
        basketDeliveryServices:
          type: array
          items:
            $ref: '#/components/schemas/ServiceModel'
        basketCostSummary:
          type: object
          properties:
            serviceSubTotalWithOutDiscount:
              type: number
            orderTotalWithDiscount:
              type: number
            lineSubTotalWithDiscount:
              type: number
            serviceSubTotalWithDiscount:
              type: number
            orderTotalWithOutDiscount:
              type: number
            cartCount:
              type: number
            cartSavings:
              type: number
            lineSubTotalWithOutDiscount:
              type: number
        products:
          type: array
          items:
            properties:
              quantity:
                type: number
              priceValues:
                type: object
                properties:
                  basePrice:
                    type: number
                  offerPrice:
                    type: number
                  lineTotal:
                    type: number  
              productId:
                type: string
              productUrl:
                type: string
              productModelId:
                type: string
              productDescription:
                type: string
              productName:
                type: string
              productServices:
                type: array
                items:
                  properties:
                    serviceType:
                      type: string
                    isMultiplier:
                      type: boolean
                    isServiceSelected:
                      type: boolean  
                    serviceDescription:
                      type: string
                    serviceId:
                      type: string
                    priceValues:
                      type: object
                      properties:
                        basePrice:
                          type: number
                        offerPrice:
                          type: number
                        lineTotal:
                          type: number  
...
    OrderModel:
      type: object
      properties:
        error:
          type: boolean
          default: false
        message:
          type: string
        OrderSummary:
          type: array
          items:
            properties:
              guid: 
                type: string
              orderId:
                type: string
              createdAt:
                type: string
              emailId:
                type: string
        OrderEntries:
          type: array
          items:
            properties:
              pk:
                type: string
              orderId:
                type: string
              entryId:
                type: string
              entryCode:
                type: string
              productModelId:
                type: string
              codeShortDescription:
                type: string
              basePrice:
                type: number
              payablePrice:
                type: number
              quantity:
                type: number
              lineTotal:
                type: number
        OrderCostSummary:
          type: array
          items:
            properties:
              orderId:
                type: string
              lineTotalWithDiscount:
                type: number
              serviceTotalWithDiscount:
                type: number
              orderTotalWithDiscount:
                type: number
              lineTotalWithoutDiscount:
                type: number
              serviceTotalWithoutDiscount:
                type: number
              orderTotalWithoutDiscount:
                type: number  
              totalQuantity:
                type: number
              orderSavings:
                type: number
        orderAddressDetails:
          type: array
          items:
            properties:
              pk:
                type: string
              orderId:
                type: string
              emailId:
                type: string
              line1:
                type: string
              line2:
                type: string
              pincode:
                type: string
              countryCode:
                type: string
              country: 
                type: string
              phoneNumber:
                type: string
              city:
                type: string
              state:
                type: string
              addressType:
                type: string
                enum:
                  - BILLING
                  - SHIPPING
              addressTitle:
                type: string
              busnisessType:
                type: string
                enum:
                  - B2C
                  - B2B
              companyName:
                type: string
    ResponseModel:
      type: object
      properties:
        error: 
          type: boolean
        message:
          type: string
...